+++ DentalSystem/DentalSystem.Entities/Requests/AccountsReceivable/AddAccountsReceivableRequest.cs
@@ -8,8 +8,8 @@ namespace DentalSystem.Entities.Requests.AccountsReceivable
     {
         public int AccountsReceivableId { get; set; }
         public int PatientId { get; set; }
-        public int Total { get; set; }
-        public int TotalPaid { get; set; }
+        public decimal Total { get; set; }
+        public decimal TotalPaid { get; set; }
         public DateTime CreatedDate { get; set; }
         public IMapper Mapper { get; set; }
         public SetVisitAsBilledRequest SetVisitAsBilledRequest { get; set; }
