+++ DentalSystem/DentalSystem.Entities/Requests/AccountsReceivable/UpdateTotalPaidRequest.cs
@@ -5,6 +5,6 @@ namespace DentalSystem.Entities.Requests.AccountsReceivable
     public class UpdateTotalPaidRequest
     {
         public int AccountsReceivableId { get; set; }
-        public int TotalPaid { get; set; }
+        public decimal TotalPaid { get; set; }
     }
 }
