+++ DentalSystem/DentalSystem/Printing/PrintingModel.cs
@@ -9,8 +9,8 @@ namespace DentalSystem.Printing
         public List<GetInvoiceDetailByVisitIdResultModel> ItemsToBill { get; set; }
         public DateTime VisitDate { get; set; }
         public string VisitNumber { get; set; }
-        public int Total { get; set; }
-        public int Paid { get; set; }
-        public int Pending { get; set; }
+        public decimal Total { get; set; }
+        public decimal Paid { get; set; }
+        public decimal Pending { get; set; }
     }
 }
